Move data config to settings module
