+++ src/machine_learning/nodes/common.py
@@ -8,9 +8,8 @@ from datetime import datetime
 import pandas as pd
 import numpy as np
 
-
-from machine_learning.data_config import INDEX_COLS, ORIGINAL_COLUMNS
-from machine_learning.settings import MELBOURNE_TIMEZONE
+from machine_learning.data_config import ORIGINAL_COLUMNS
+from machine_learning.settings import MELBOURNE_TIMEZONE, INDEX_COLS
 from .base import _validate_required_columns
